requirements.txt
pyTelegramBotAPI
python-dotenv
requests
flake8